Add the clipboard script to templates
Added the script tags pointing to clipboard.js to the index and project
html files, allowing us to use it's features to add the clone urls to
the users clipboard, without asking them to highlight and ctrl-c. This
feature does require the user to click a button, we are not auto adding
data to their clipboard.