Change directory icon from octicons to fontawesome
Change the directory icons from octicons to fontawesome. This allows us
to have an open and a closed icon. Also changed how the icon is placed.
Instead of using an image tag we now use the CSS3 :before pseudoclass,
which allows us to automatically switch between the open and closed
folder.